+++ workers/email/src/templates/emails/assignee/html.twig
@@ -15,21 +15,12 @@
 					{% else %}
 						{{ whoAssigned.email }}
 					{% endif %}
-					assigned you to the issue
+					assigned you to resolve the event
 				</span>
 			</font>
 		</td>
 	</tr>
-	<tr>
-		<td style="display: block; padding-top: 20px; padding-right: 20px; padding-bottom: 20px; padding-left: 20px; margin-bottom: 60px; border-width: 1px; border-color: #494f5e; border-style: solid; border-radius: 10px;">
-			<font color="#dbe6ff" style="font-size: 15px; letter-spacing: 0.4px;">
-				<span style="display: block; padding-bottom: 15px;">
-					{{ event.payload.title }}
-				</span>
-			</font>
-			{% include '../../components/event-info.twig' with {event: { daysRepeated: daysRepeated, event: {totalCount: event.totalCount}, usersAffected: event.usersAffected}} %}
-		</td>
-	</tr>
+	{% include '../../components/issue.twig' with daysRepeated, event %}
 	<tr>
 		<td style="padding-right: 20px;padding-left: 20px;padding-bottom: 40px;">
 			{% set url = host ~ '/project/' ~ project._id ~ '/event/' ~ event._id ~ '/overview' %}
