+++ workers/email/src/templates/emails/assignee/text.twig
@@ -1,4 +1,4 @@
-{% if whoAssigned.name %}{{ whoAssigned.name }}{% else %}{{ whoAssigned.email }}{% endif %} assigned you to the issue
+{% if whoAssigned.name %}{{ whoAssigned.name }}{% else %}{{ whoAssigned.email }}{% endif %} assigned you to resolve the event
 
 {{ event.payload.title }}
