+++ workers/sender/src/provider.ts
@@ -1,5 +1,5 @@
 import { Logger } from 'winston';
-import { AllNotifications } from '../types/template-variables';
+import { Notification } from 'hawk-worker-sender/types/template-variables/';
 
 /**
  * Abstract class that describes notification provider
@@ -27,5 +27,5 @@ export default abstract class NotificationsProvider {
    * @param endpoint - endpoint where to send notification
    * @param variables - notification variables
    */
-  public abstract async send(endpoint: string, variables: AllNotifications): Promise<void>;
+  public abstract async send(endpoint: string, variables: Notification): Promise<void>;
 }
