+++ workers/sender/tests/provider-example.ts
@@ -1,6 +1,6 @@
 /* eslint-disable */
 import NotificationsProvider from '../src/provider';
-import { AllNotifications } from '../types/template-variables';
+import { Notification } from '../types/template-variables';
 
 /**
  * Example implementation of abstract provider
@@ -10,9 +10,9 @@ export default class ExampleProvider extends NotificationsProvider {
    * Send method example
    *
    * @param to - endpoint
-   * @param variables - template variables
+   * @param notification - notification to send
    */
-  public async send(to: string, variables: AllNotifications): Promise<void> {
+  public async send(to: string, notification: Notification): Promise<void> {
     // console.log(`ExampleProvider: sending to ${to} ... `);
     // console.log('With variables', variables);
   }
