+++ workers/sender/types/template-variables/common-template.ts
@@ -1,0 +1,14 @@
+/**
+ * Common variables for notification templates
+ */
+export interface CommonTemplateVariables {
+  /**
+   * Api url
+   */
+  host: string;
+
+  /**
+   * Api url with path to static files
+   */
+  hostOfStatic: string;
+}
