+++ workers/sender/types/template-variables/notification.ts
@@ -1,0 +1,14 @@
+/**
+ * Notification with payload and type
+ */
+export interface Notification<Payload> {
+  /**
+   * Notification type
+   */
+  type: string;
+
+  /**
+   * Notification payload
+   */
+  payload: Payload;
+}
