+++ workers/slack/src/templates/new-event.ts
@@ -1,4 +1,4 @@
-import type { EventsTemplateVariables, TemplateEventData } from 'hawk-worker-sender/types/template-variables';
+import type { EventsTemplateVariables, TemplateEventData } from 'hawk-worker-sender/types/template-variables/';
 import { IncomingWebhookSendArguments } from '@slack/webhook';
 import { block, element, object, TEXT_FORMAT_MRKDWN, TEXT_FORMAT_PLAIN } from 'slack-block-kit';
 import { GroupedEventDBScheme } from 'hawk.types';
