+++ workers/slack/tests/__mocks__/several-events-notify.ts
@@ -1,4 +1,4 @@
-import { EventsTemplateVariables, SeveralEventsNotification } from 'hawk-worker-sender/types/template-variables';
+import { EventsTemplateVariables, SeveralEventsNotification } from 'hawk-worker-sender/types/template-variables/';
 import { GroupedEventDBScheme, ProjectDBScheme } from 'hawk.types';
 import { ObjectId } from 'mongodb';
