+++ workers/telegram/src/provider.ts
@@ -1,4 +1,4 @@
-import { EventsTemplateVariables, AllNotifications } from 'hawk-worker-sender/types/template-variables';
+import { EventsTemplateVariables, Notification } from 'hawk-worker-sender/types/template-variables/';
 import NotificationsProvider from 'hawk-worker-sender/src/provider';
 import templates from './templates';
 import { TelegramTemplate } from '../types/template';
@@ -12,21 +12,18 @@ export default class TelegramProvider extends NotificationsProvider {
    * Send telegram message to recipient
    *
    * @param to - recipient endpoint
-   * @param variables - variables for template
+   * @param notification - notification with payload and type
    */
-  public async send(to: string, variables: AllNotifications): Promise<void> {
+  public async send(to: string, notification: Notification): Promise<void> {
     let template: TelegramTemplate;
 
-    if (variables.type === 'event') {
-      template = templates.NewEventTpl;
-    } else if (variables.type === 'several-events') {
-      template = templates.SeveralEventsTpl;
-    } else {
+    switch (notification.type) {
+      case 'event': template = templates.NewEventTpl; break;
+      case 'several-events':template = templates.SeveralEventsTpl; break;
       // todo: add assignee notification
-      return;
     }
 
-    const message = await this.render(template, variables.payload as EventsTemplateVariables);
+    const message = await this.render(template, notification.payload as EventsTemplateVariables);
 
     await this.deliver(to, message);
   }
