+++ workers/telegram/tests/provider.test.ts
@@ -1,4 +1,4 @@
-import { EventsTemplateVariables, EventNotification, SeveralEventsNotification } from 'hawk-worker-sender/types/template-variables';
+import { EventsTemplateVariables, EventNotification, SeveralEventsNotification } from 'hawk-worker-sender/types/template-variables/';
 import { DecodedGroupedEvent, ProjectDBScheme } from 'hawk.types';
 import TelegramProvider from 'hawk-worker-telegram/src/provider';
 import templates from '../src/templates';
